login: add support for shadow passwords
login, on fallback from PAM, or when PAM support is not enabled, checks
pw->pw_passwd for locked ("!") or passwordless ("*") accounts.  However,
on systems with shadow passwords the first character will always be "x".

This patch adds shadow password support from the passwd tool, letting
the user end up in "Login incorrect" rather than the "login: bad salt"
case, which could be used by an attacker to guess the state of accounts.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>